refactoring + minor style fixes
