+++ src/components/organisms/WithNavigationBar/WithNavigationBar.tsx
@@ -20,10 +20,12 @@ const NavBar = lazy(() =>
 
 export interface WithNavigationBarProps {
   hasBackButton?: boolean;
+  withSchedule?: boolean;
 }
 
 export const WithNavigationBar: React.FC<WithNavigationBarProps> = ({
   hasBackButton,
+  withSchedule = true,
   children,
 }) => {
   // @debt remove useVenueId from here and just pass it through as a prop/similar
@@ -39,7 +41,7 @@ export const WithNavigationBar: React.FC<WithNavigationBarProps> = ({
        */}
       <RelatedVenuesProvider venueId={venueId}>
         <Suspense fallback={<Loading />}>
-          <NavBar hasBackButton={hasBackButton} />
+          <NavBar hasBackButton={hasBackButton} withSchedule={withSchedule} />
         </Suspense>
       </RelatedVenuesProvider>
