+++ src/pages/EmergencyViewPage/EmergencyViewPage.scss
@@ -1,228 +1,40 @@
 @import "scss/constants.scss";
+@import "../VenueLandingPage/VenueLandingPage.scss";
 
+$tab-item-height: 35px;
+$tab-item-image-height: 70px;
+$tab-item-image-width: 100px;
+$tab-item-room-height: 170px;
+$tab-item-room-width: 180px;
+
+// hide landing page content
 .venue-entrance-experience-container {
   height: 0px;
-
-  .count-down-container {
-    font-size: 15px;
-  }
-
-  .playa-container {
-    overflow: hidden;
-    position: relative;
-
-    .playa-abs-container {
-      position: absolute;
-      top: 0;
-      left: 0;
-      bottom: 0;
-      right: 0;
-      display: flex;
-      justify-content: center;
-      align-items: center;
-
-      .playa-icon {
-        width: 150px;
-        height: 150px;
-        object-fit: cover;
-        border-radius: 28px;
-        overflow: hidden;
-      }
-
-      .playa-icon-container {
-        position: relative;
-      }
-
-      .playa-venue-preview {
-        position: absolute;
-        top: -200px;
-        right: -600px;
-        width: 550px;
-        height: 600px;
-        background-color: black;
-        overflow: auto;
-        border-radius: 20px;
-      }
-    }
-  }
-
-  .login-invitation {
-    font-size: 14px;
-    margin-top: 8px;
-  }
-  .quotation-container {
-    display: flex;
-    flex-direction: column;
-    color: $white;
-    align-items: center;
-    margin-bottom: 10px;
-
-    .quotation {
-      font-weight: 500;
-      font-size: 18px;
-      margin-bottom: 5px;
-      text-align: center;
-    }
-    .quotation-author {
-      font-size: 13px;
-    }
-  }
-
-  .toast {
-    position: absolute;
-    max-width: none;
-    width: 40%;
-    margin: auto;
-    background: $light-green;
-    color: $black;
-
-    .toast-header {
-      background: $dark-green;
-      color: $black;
-      display: flex;
-      justify-content: space-between;
-
-      button {
-        color: $black;
-        opacity: 0.7;
-        margin: 0;
-        outline: none;
-        border: none;
-
-        &:focus {
-          outline: none;
-          border: none;
-        }
-      }
-    }
-  }
-
-  .toast-container {
-    position: absolute;
-    width: 100%;
-    top: 60px;
-    right: 0px;
-    width: 100%;
-    display: flex;
-    justify-content: center;
-  }
-  .header {
-    display: flex;
-    justify-content: space-between;
-    align-items: flex-end;
-    object-fit: cover;
-
-    .secret-password-form-wrapper {
-      background: rgba($black, 0.75);
-      padding: 20px;
-      width: 250px;
-      margin-right: 40px;
-      margin-bottom: 30px;
-      border-radius: 22px;
-    }
-
-    .venue-host {
-      margin: 30px 50px;
-
-      .title {
-        font-size: 50px;
-      }
-
-      .subtitle {
-        font-size: 20px;
-      }
-    }
-
-    .countdown {
-      font-weight: 400;
-      display: block;
-      font-style: italic;
-    }
-  }
-
-  .venue-presentation {
-    padding: 20px;
-
-    .presentation-paragraph {
-      color: $concrete;
-    }
-
-    .check-icon-container {
-      margin-right: 10px;
-      color: $primaryColor;
-      font-size: 25px;
-    }
-
-    .checklist-item {
-      margin: 20px 0;
-      display: flex;
-      font-size: 18px;
-      align-items: center;
-      font-weight: 500;
-    }
-
-    .youtube-video {
-      border-radius: 22px;
-      margin-bottom: 20px;
-    }
-  }
-}
-
-.oncoming-events {
-  .upcoming-gigs-title {
-    margin: 20px 0;
-    font-size: 25px;
-    font-weight: 700;
-  }
-  h4 {
-    font-size: 30px;
-    font-style: italic;
-  }
-  .date {
-    font-size: 15px;
-    color: $white;
-    margin-bottom: 5px;
-  }
-  .event-description {
-    margin-bottom: 20px;
-    font-size: 15px;
-  }
-
-  .button-container {
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-  }
-  .disabled {
-    background: rgba($dark, 0.6);
-  }
-  .information-card-container {
-    margin-bottom: 20px;
-    &:last-child {
-      margin-bottom: 50px;
-    }
-  }
 }
 
 .EmergencyView {
   width: 100%;
-  height: 100%;
+  height: auto;
   background-color: opaque-black(0.8);
-  position: absolute;
+  position: relative;
   display: flex;
   flex-direction: column;
-  z-index: 100;
+  z-index: z(emergency-content);
+
+  &_main {
+    margin-bottom: $spacing--xxl;
+  }
 
   &__weekdays {
     display: flex;
     flex-wrap: wrap;
-    margin: 30px;
+    margin: $spacing--xxl;
 
     &__column {
       width: 33%;
       display: flex;
       flex-direction: column;
-      padding-left: 90px;
+      padding-left: $spacing--xxl * 3;
       cursor: default;
 
       .ScheduleItemNG {
@@ -236,51 +48,52 @@
     margin: 0 auto;
     width: 15%;
     justify-content: space-around;
-    font-size: 20px;
+    font-size: $font-size--xl;
     font-weight: 900;
-    margin-top: 22px;
+    margin-top: $spacing--xl;
     cursor: pointer;
 
     &_item {
-      height: 35px;
+      height: $tab-item-height;
 
       &--active {
-        border-bottom: 3px solid $primary--live;
+        border-bottom: $border-radius--xs solid $primary--live;
       }
     }
   }
 
   &_venue {
-    margin: 14px;
+    margin: $spacing--lg;
     font-weight: 900;
-    font-size: 16px;
+    font-size: $font-size--lg;
   }
 
   &_content {
     display: flex;
 
     &_room {
-      height: 170px;
-      width: 180px;
-      border-radius: 12px;
+      height: $tab-item-room-height;
+      width: $tab-item-room-width;
+      border-radius: $border-radius--md;
       background-color: opaque-white(0.2);
       display: flex;
       flex-direction: column;
       align-items: center;
-      margin: 14px;
+      margin: $spacing--lg;
+      cursor: default;
 
       &_info {
         display: flex;
         width: 90%;
-        margin-top: 8px;
+        margin-top: $spacing--sm;
 
         &_status {
           display: flex;
           background-color: $primary--live;
-          border-radius: 12px;
-          padding: 8px;
-          width: 60px;
-          height: 30px;
+          border-radius: $border-radius--md;
+          padding: $spacing--sm;
+          width: $spacing--xxl * 2;
+          height: $spacing--xxl;
           align-items: center;
           justify-content: center;
         }
@@ -288,14 +101,15 @@
         &_audience {
           display: flex;
           background-color: opaque-white(0.2);
-          border-radius: 12px;
-          padding: 8px;
-          width: 60px;
-          height: 30px;
+          border-radius: $border-radius--md;
+          padding: $spacing--sm;
+          width: $spacing--xxl * 2;
+          height: $spacing--xxl;
           justify-content: space-around;
           margin-left: auto;
+
           span {
-            font-size: 12px;
+            font-size: $font-size--sm;
           }
         }
       }
@@ -305,11 +119,14 @@
         display: flex;
         flex-direction: column;
         align-items: center;
+        cursor: pointer;
+        justify-content: space-between;
+        height: 60%;
 
         &_image {
-          width: 100px;
-          height: 70px;
-          padding: 4px;
+          width: $tab-item-image-width;
+          height: $tab-item-image-height;
+          padding: $spacing--xs;
         }
       }
     }
