+++ src/pages/EmergencyViewPage/EmergencyViewPageRooms.tsx
@@ -1,0 +1,52 @@
+import React from "react";
+
+import { AnyVenue, ScheduledVenueEvent } from "types/venues";
+
+import { isEventLive } from "utils/event";
+import { WithId } from "utils/id";
+
+import EmergencyViewRoom from "./EmergencyViewRoom";
+
+type EmergencyViewTabsProps = {
+  descendantVenues: WithId<AnyVenue>[];
+  liveAndFutureEvents: ScheduledVenueEvent[];
+};
+
+const EmergencyViewPageRooms: React.FC<EmergencyViewTabsProps> = ({
+  descendantVenues,
+  liveAndFutureEvents,
+}) => {
+  const liveEvents = liveAndFutureEvents.filter(isEventLive);
+
+  const isRoomHasLiveEvent = (room: string) =>
+    liveEvents.map((event) => event.room).includes(room);
+
+  const renderVenuesRooms = descendantVenues
+    .map(
+      (venue) =>
+        !!venue?.rooms?.length && (
+          <>
+            <span className="EmergencyView_venue">{venue.id}</span>
+            <div key={venue.id} className="EmergencyView_content">
+              {venue?.rooms?.map((room) => {
+                const isRoomLive = isRoomHasLiveEvent(room.title);
+
+                return (
+                  <EmergencyViewRoom
+                    key={room.title}
+                    room={room}
+                    venueName={venue.name}
+                    isLive={isRoomLive}
+                  />
+                );
+              })}
+            </div>
+          </>
+        )
+    )
+    .filter((venue) => !!venue);
+
+  return <>{renderVenuesRooms}</>;
+};
+
+export default EmergencyViewPageRooms;
